# src/main/resources/api-definition.yml
swagger: '2.0'
info:
  description: Server to manage events for Beevy
  version: 1.0.0
  title: Beevy App
...
parameters:
  eventID:
    name: eventID
    in: path
    description: ID of a User
    required: true
    type: string
  userID:
    name: userID
    in: path
    description: ID of a User
    required: true
    type: string
  username:
    name: username
    in: path
    description: username
    required: true
    type: string
  tempAccessToken:
    name: tempAccessToken
    in: path
    description: tempAccessToken
    required: true
    type: string
paths:
  /user/register:
    post:
      tags:
        - user
      summary: Register user
      description: Register a user via mail
      operationId: registerUser
      consumes:
          - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: Minimal User
          schema:
            $ref: '#/definitions/User'
      responses:
        '201':
          description: Sent registration mail
        '405':
          description: Failed to send registration mail
  /user/avatar:
    post:
      tags:
        - user
      summary: Update user avatar
      description: Set a new user avatar
      operationId: updateAvatar
      consumes:
          - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: AvatarDTO
          schema:
            $ref: '#/definitions/AvatarDTO'
      responses:
        '201':
          description: Updated user avatar
        '405':
          description: Failed to update avatar
  /user/access:
    post:
      tags:
        - user
      summary: Set a temporary access token for a user
      description: Set a temporary access token to make sure only the right user can get data
      operationId: setTempAccessTokenForUser
      consumes:
          - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: Security Object
          schema:
            $ref: '#/definitions/UserSecurity'
      responses:
        '201':
          description: Saved temporary access token
        '405':
          description: Failed to create temporary access token
  /user/{username}/{userID}/{tempAccessToken}:
    parameters:
      - $ref: '#/parameters/username'
      - $ref: '#/parameters/userID'
      - $ref: '#/parameters/tempAccessToken'
    get:
      tags:
        - user
      summary: Return security token for a user
      description: Get security token
      operationId: getUserToken
      produces:
        - application/json
      responses:
        '200':
          description: successful operation
          schema:
            type: string
        '404':
          description: User not found
...
  /comment:
    post:
      tags:
        - comment
      summary: Add a comment
      description: Add a comment to an event or another comment
      operationId: addComment
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: Comment Data
          schema:
            $ref: '#/definitions/CommentDTO'
      responses:
        '201':
          description: Added comment
        '405':
          description: Failed to add comment
  /comment/{eventID}/{userID}/{tempAccessToken}:
    parameters:
      - $ref: '#/parameters/eventID'
      - $ref: '#/parameters/userID'
      - $ref: '#/parameters/tempAccessToken'
    get:
      tags:
        - comment
      summary: Get comments on an event
      description: Load all comments on an eventr
      operationId: getComments
      produces:
        - application/json
      responses:
        '200':
          description: successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/Comment'
        '405':
          description: Failed to get comments
definitions:
...
  AvatarDTO:
    type: object
    properties:
      userID:
        type: string
      token:
        type: string
      avatar:
        type: string
...
  User:
    type: object
    properties:
      username:
        type: string
      userID:
        type: string
      mail:
        type: string
      token:
        type: string
      tempAccessToken:
        type: string
      joinedEvents:
        type: array
        items:
          type: string
      createdEvents:
        type: array
        items:
          type: string
  UserSecurity:
    type: object
    properties:
      username:
        type: string
      userID:
        type: string
      token:
        type: string
      tempToken:
        type: string
...
  CommentDTO:
    type: object
    properties:
      userID:
        type: string
      userToken:
        type: string
      eventID:
        type: string
      repliedTo:
        type: string
      commentBody:
        type: string
      commentTime:
        type: string
  Comment:
    type: object
    properties:
      commentID:
        type: string
      author:
        type: string
      authorID:
        type: string
      createdAt:
        type: string
      commentBody:
        type: string
      comments:
        type: array
        items:
          $ref: '#/definitions/Comment'